+++ baseimage/setup/apps/etc/skel/clone/build/install.sh.tpl
@@ -7,10 +7,13 @@ rm -rf /apps; mkdir /apps
 
 # copy everything from setup to the root /apps except Dockerfile rebuild materials
 echo copying application files to /apps ...
-tar cf - --exclude ./build --exclude ./build.sh --exclude ./run.sh . | (cd /apps; tar xf -)
+tar cf - \
+   --exclude ./build --exclude ./build.sh --exclude Dockerfile \
+   --exclude ./run.sh --exclude '*~' --exclude 'var/*' . \
+   | (cd /apps; tar xf -)
 
-# update the version information
-mv /setup/build/new_version.inc /apps/etc/version.inc
+# Update the version information, if a replacement exists
+[ -f /setup/build/new_version.inc ] && mv /setup/build/new_version.inc /apps/etc/version.inc
 
 # Add additional setup commands for your production image here, if any.
 # ...
